housekeeping move stargazer tools
